c2go,tests,docs: I-C corrective pass — isolated parity for every multipolygon symbol
Review round for #34 (reopened):

1. Symbol-by-symbol parity matrix (posted on #34): every I-C-owned
   function classified direct / composite-only / justified n/a. The
   previously composite-only functions now have isolated C<->Go
   comparisons via new same-TU wrappers:
   - cellToEdgeArcs (per-arc ids/linkage/parent/rank, hexagon +
     pentagon idx tables), createArcSet hash-bucket layout,
     resetVisited (visited flags before/after), unionArcs/getRoot
     (scripted pair sequences incl. rank-swap and a==b branches, roots
     AND ranks), createSortableLoop (per surviving arc),
     cmp_SortableLoop/cmp_SortablePoly/cmp_uint64 (all branches),
     createSortablePoly (0 and 1 holes), createMultiPolygon (isolated
     from validate/overflow; globe branch), destroyArcSet/
     destroySortableLoopSet(+Shallow)/destroyGeoLoop/destroyGeoPolygon/
     destroyGeoMultiPolygon (nulled-field state + double-call safety);
   - linkedGeo conversion statics via guarded wrappers in
     h3lib_linkedGeo_c2go.c: linkedGeoLoopToGeoLoop,
     geoLoopToLinkedGeoLoop, linkedGeoPolygonToGeoPolygon,
     addLinkedGeoLoop (first + append), geoPolygonToLinkedGeoLoops,
     the isolated geoMultiPolygonToLinkedGeoPolygon success path, and
     the conversion externs' error branches (scripted upstream
     constructions, codes exact).
   destroySortablePolys/destroySortablePolyVerts are recorded as
   justified n/a (C frees the array; freed memory cannot be legally
   inspected) with real ordinary Go assertions instead.

2. Dedicated destroyLinkedMultiPolygon idempotence parity regression
   (4.5.0 delta): both sides destroy twice over all 10 input sets;
   neither second call may fail; head zeroed after both calls. C side
   h3v450-gated (4.4.0 C had no such contract).

3. destroySortablePolys/destroySortablePolyVerts ordinary tests now
   assert the nested Holes/GeoLoop references are cleared; the
   implementation comments state honestly that a by-value slice
   header cannot be nilled caller-visibly.

4. docs/lint-policy.md: nolint:unused count corrected to 16, deleted
   file names replaced, area_geoMultiPolygonAreaRads2.go added to the
   unparam list.

5. checkCellsToMultiPolyOverflow: cSizeMax is now the platform SIZE_MAX
   equivalent (^uint(0)); the pinned LP64 element sizes are documented
   as safe overestimates on 32-bit targets (strictly more conservative
   than a 32-bit C build) — bit-for-bit oracle parity is claimed for
   64-bit targets only, where the parity suite runs.

6. CellsToMultiPolygon allocation acceptance: GoDoc documents that the
   call allocates per invocation with no contractual bound (count
   scales with input); TestWave3Allocations pins measured budgets for
   fixed inputs (8 single-cell, 15 donut) as regression tripwires.

+++ cellsToMultiPoly__checkCellsToMultiPolyOverflow.go
@@ -1,14 +1,19 @@
 package h3
 
-// C sizes on the LP64 platforms the parity oracle runs on: sizeof(Arc)
-// = 48 (8 id + 2 bools + 6 padding + 3×8 pointers + 8 rank), a pointer
-// is 8 bytes, and SIZE_MAX = 2^64-1. The Go port pins these so the
-// overflow threshold matches the C oracle bit-for-bit (verified by the
-// parity suite); Go's own allocator limits are unrelated.
+// Element sizes for the overflow pre-check. The struct sizes are the
+// LP64 C oracle's (sizeof(Arc) = 48: 8 id + 2 bools + 6 padding + 3×8
+// pointers + 8 rank; 8-byte pointers); on 32-bit Go targets the real
+// element sizes are smaller, so using these values makes the check
+// strictly more conservative there (it rejects earlier than a 32-bit C
+// build would), never less safe. cSizeMax is the platform's SIZE_MAX
+// equivalent (^uint(0)): 2^64-1 on 64-bit targets — where it reproduces
+// the C oracle's threshold bit-for-bit (the parity suite runs on
+// 64-bit platforms only) — and 2^32-1 on 32-bit targets, where the
+// stricter threshold guards Go's own allocation-length limits.
 const (
 	cSizeofArc    = 48
 	cSizeofArcPtr = 8
-	cSizeMax      = ^uint64(0)
+	cSizeMax      = uint64(^uint(0))
 )
 
 // checkCellsToMultiPolyOverflow checks for potential integer overflow in
@@ -33,7 +38,7 @@ func checkCellsToMultiPolyOverflow(numCells int64, hashMultiplier int64) h3Error
 
 	// Check if maxBytesPerCell * numCells would overflow size_t, which is
 	// what is used for allocations. Use SIZE_MAX since size_t may be 32
-	// bits (the oracle platforms are 64-bit; see cSizeMax above).
+	// bits (cSizeMax mirrors the platform width; see above).
 	if numCells > 0 && uint64(numCells) > cSizeMax/maxBytesPerCell {
 		return eMemoryBounds
 	}
